+++ CHANGELOG.md
@@ -1,8 +1,16 @@
 # Changelog
 
+## 0.1.1
+
+### Updated
+
+- More perfect typing added
+- Made the worker config optional
+- Updated readme with the docs
+
 ## 0.1.0
 
-**ADDED**
+### ADDED
 
 - Worker name specification.
 - Worker priority set by score.